+++ src/eval_ab_3d_mot/cli/batch_ab_3d_mot_annotations/cmd_line_factory.py
@@ -7,18 +7,18 @@ from pure_ab_3d_mot.dist_metrics import MetricKind
 from pure_ab_3d_mot.matching import MatchingAlgorithm
 from rich_argparse import RawTextRichHelpFormatter
 
-from eval_ab_3d_mot.cli.batch_ab_3d_mot.cmd_line_object import CmdLineBatchRunAb3dMot
 from eval_ab_3d_mot.cli.common.get_hlp import get_hlp
 from eval_ab_3d_mot.cli.common.init_coinciding_attrs import init_coinciding_attrs
-from eval_ab_3d_mot.cli.common.kitti_category import CATEGORIES, HLP_CATEGORY
 from eval_ab_3d_mot.cli.common.tracker_meta import AUTO
 from eval_ab_3d_mot.kitti_category import KittiCategory
 
+from .cmd_line_object import CmdLineBatchRunAb3dMotAnnotations
 
-PROG = 'batch-run-ab-3d-mot'
+
+PROG = 'batch-run-ab-3d-mot-annotations'
 HLP_OUT = 'Directory to store tracking results.'
-HLP_ANN = 'Annotations (ground-truth) directory.'
 DEF_POLICY = 'If not given the category-dependent optimal will be used.'
+HLP_CAT_OBJ = 'Category of the objects selected for tracking.'
 HLP_THR = 'Association threshold.'
 HLP_ALG = 'Association algorithm.'
 HLP_MET = 'Association metric.'
@@ -26,16 +26,16 @@ HLP_MAX = 'Maximal number of steps without association.'
 HLP_CAT_PRM = 'Category of to selected tracker parameters.'
 
 
-def get_cmd_line(args: Sequence[str]) -> CmdLineBatchRunAb3dMot:
-    cli = CmdLineBatchRunAb3dMot()
+def get_cmd_line(args: Sequence[str]) -> CmdLineBatchRunAb3dMotAnnotations:
+    cli = CmdLineBatchRunAb3dMotAnnotations()
     parser = ArgumentParser(
-        PROG, f'{PROG} <detections+> [OPTIONS]', formatter_class=RawTextRichHelpFormatter
+        PROG, f'{PROG} <annotations+> [OPTIONS]', formatter_class=RawTextRichHelpFormatter
     )
-    parser.add_argument('detections', nargs='+', help='Detection files.')
-    parser.add_argument('--ann-dir', '-ad', help=get_hlp(HLP_ANN, cli.ann_dir))
+    parser.add_argument('annotations', nargs='+', help='KITTI annotation files.')
     parser.add_argument('--trk-dir', '-o', help=get_hlp(HLP_OUT, cli.trk_dir))
-    hlp_category = get_hlp(HLP_CATEGORY, cli.category_obj)
-    parser.add_argument('--category-obj', '-c', choices=CATEGORIES, help=hlp_category)
+    cc_oo = tuple(c.value for c in KittiCategory)
+    hlp_c_obj = get_hlp(HLP_CAT_OBJ, cli.category_obj)
+    parser.add_argument('--category-obj', '-c', choices=cc_oo, help=hlp_c_obj)
     hlp_c_prm = get_hlp(HLP_CAT_PRM, cli.category_prm + ' 🛈 objects category')
     cc_pp = tuple(c.value for c in KittiCategory) + (AUTO,)
     parser.add_argument('--category-prm', '-p', choices=cc_pp, help=hlp_c_prm)