+++ src/eval_ab_3d_mot/cli/common/single_sequence.py
@@ -6,11 +6,12 @@ import numpy as np
 
 from pure_ab_3d_mot.tracker import Ab3DMot
 
+from eval_ab_3d_mot.cli.common.kitti_adaptor import KittiAdaptor
 from eval_ab_3d_mot.cli.common.r_cnn_adaptor import DETS, RCnnAdaptor
 
 
 def get_tracking_result(
-    adaptor: RCnnAdaptor, tracker: Ab3DMot, verbosity: int
+    adaptor: RCnnAdaptor | KittiAdaptor, tracker: Ab3DMot, verbosity: int
 ) -> List[List[np.ndarray]]:
     result = []
     for step, det_dct in enumerate(adaptor.detections_3d()):