+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Possibility to report on the tracker parameters to stdout in the batch run script.
   - The number of figures after floating point is 4 for F1, Precision and Recall and 6 for Accuracy.
   - Report tracker parameters once, before tracking in `run-ab-3d-mot-with-clavia` and `batch-run-ab-3d-mot`.
+  - Script `batch-run-ab-3d-mot-annotations` to run the tracker consuming KITTI *annotations*. 
 
 # 2.1.0